pyproject.toml
[project]
name = "websocket-control-stackchan-server"
version = "0.1.0"
description = "A WebSocket control interface for StackChan AI agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[websockets]>=0.128.0",
    "google-cloud-speech>=2.35.0",
    "uvicorn[standard]>=0.40.0",
    "voicevox-client>=1.1.0",
]

[dependency-groups]
sample-gemini = [
    "google-genai>=1.59.0",
]
